feat: Implement line counting functionality and update build process to output line count to count.txt
