+++ Makefile
@@ -6,7 +6,7 @@ VERSION=0.2.1
 # Windows: Download from https://github.com/upx/upx/releases
 
 build:
-	go run toolkit/count.go
+	go run counter.go > toolkit/count.txt
 	make build_android
 	make build_linux
 	make build_windows
